Real signals are now float. New representation of recursive signal expressions. Big reorganisation of the typing system (it is now possible to use getType without environment parameter). Added getSubSignal function for easier signal traversal. Recursivness analysis added to prepare future modifications in code generation.

+++ compiler/signals/signals.cpp
@@ -19,28 +19,28 @@
  ************************************************************************
  ************************************************************************/
 
- 
+
 /*****************************************************************************
 	HISTORY
 	22/01/05 : added int cast in se"lect2 and select3 selection signal
 *****************************************************************************/
- 
- 
- 
+
+
+
 #include "signals.hh"
 
 
 ////////////////////////////////////////////////////////////////////////
 /**
  * Les Signaux.
- * Les différents constructeurs et deconstructeurs de signaux 
+ * Les différents constructeurs et deconstructeurs de signaux
  */
 ///////////////////////////////////////////////////////////////////////
 
 Tree  sigInt(int i) 								{ return tree(i); 					}
 bool  isSigInt(Tree t, int* i) 						{ return isInt(t->node(), i); 		}
 
-Tree  sigReal(double r) 							{ return tree(r); 					}
+Tree  sigReal(float r) 							{ return tree(r); 					}
 bool  isSigReal(Tree t, float* r) 					{ return isFloat(t->node(), r); 	}
 
 Sym SIGINPUT = symbol ("sigInput");
@@ -130,9 +130,9 @@ bool isProj (Tree t, int* i, Tree& rgroup)		{ Tree x; return isTree(t, SIGPROJ, 
 Sym SIGINTCAST = symbol ("sigIntCast");
 Sym SIGFLOATCAST = symbol ("sigFloatCast");
 
-Tree  sigIntCast(Tree t)						{ return tree(SIGINTCAST, t); 			}				
+Tree  sigIntCast(Tree t)						{ return tree(SIGINTCAST, t); 			}
 Tree  sigFloatCast(Tree t)						{ return tree(SIGFLOATCAST, t); 		}
-			
+
 bool  isSigIntCast(Tree t, Tree& x)				{ return isTree(t, SIGINTCAST, x); 		}
 bool  isSigFloatCast(Tree t, Tree& x)			{ return isTree(t, SIGFLOATCAST, x); 	}
 
@@ -156,13 +156,13 @@ bool isSigCheckbox (Tree s, Tree& lbl)		{ return isTree(s, SIGCHECKBOX, lbl);			
 
 
 Sym SIGHSLIDER = symbol ("SigHSlider");
-Tree sigHSlider   (Tree lbl, Tree cur, Tree min, Tree max, Tree step)				
+Tree sigHSlider   (Tree lbl, Tree cur, Tree min, Tree max, Tree step)
 											{ return tree(SIGHSLIDER, lbl, list4(cur,min,max,step));		}
 bool isSigHSlider (Tree s)					{ Tree lbl, params; return isTree(s, SIGHSLIDER, lbl, params);	}
 
-bool isSigHSlider (Tree s, Tree& lbl, Tree& cur, Tree& min, Tree& max, Tree& step)		
-{ 
-	Tree params; 
+bool isSigHSlider (Tree s, Tree& lbl, Tree& cur, Tree& min, Tree& max, Tree& step)
+{
+	Tree params;
 	if (isTree(s, SIGHSLIDER, lbl, params)) {
 		cur = nth(params, 0);
 		min = nth(params, 1);
@@ -176,13 +176,13 @@ bool isSigHSlider (Tree s, Tree& lbl, Tree& cur, Tree& min, Tree& max, Tree& ste
 
 
 Sym SIGVSLIDER = symbol ("SigVSlider");
-Tree sigVSlider   (Tree lbl, Tree cur, Tree min, Tree max, Tree step)				
+Tree sigVSlider   (Tree lbl, Tree cur, Tree min, Tree max, Tree step)
 											{ return tree(SIGVSLIDER, lbl, list4(cur,min,max,step));		}
 bool isSigVSlider (Tree s)					{ Tree lbl, params; return isTree(s, SIGVSLIDER, lbl, params);	}
 
-bool isSigVSlider (Tree s, Tree& lbl, Tree& cur, Tree& min, Tree& max, Tree& step)		
-{ 
-	Tree params; 
+bool isSigVSlider (Tree s, Tree& lbl, Tree& cur, Tree& min, Tree& max, Tree& step)
+{
+	Tree params;
 	if (isTree(s, SIGVSLIDER, lbl, params)) {
 		cur = nth(params, 0);
 		min = nth(params, 1);
@@ -196,13 +196,13 @@ bool isSigVSlider (Tree s, Tree& lbl, Tree& cur, Tree& min, Tree& max, Tree& ste
 
 
 Sym SIGNUMENTRY = symbol ("SigNumEntry");
-Tree sigNumEntry   (Tree lbl, Tree cur, Tree min, Tree max, Tree step)				
+Tree sigNumEntry   (Tree lbl, Tree cur, Tree min, Tree max, Tree step)
 											{ return tree(SIGNUMENTRY, lbl, list4(cur,min,max,step));		}
 bool isSigNumEntry (Tree s)					{ Tree lbl, params; return isTree(s, SIGNUMENTRY, lbl, params);	}
 
-bool isSigNumEntry (Tree s, Tree& lbl, Tree& cur, Tree& min, Tree& max, Tree& step)		
-{ 
-	Tree params; 
+bool isSigNumEntry (Tree s, Tree& lbl, Tree& cur, Tree& min, Tree& max, Tree& step)
+{
+	Tree params;
 	if (isTree(s, SIGNUMENTRY, lbl, params)) {
 		cur = nth(params, 0);
 		min = nth(params, 1);
@@ -234,7 +234,7 @@ Sym SIGATTACH = symbol ("sigAttach");
 Tree  sigAttach(Tree t0, Tree t1)					{ return tree(SIGATTACH, t0, t1); 		}
 bool  isSigAttach(Tree t, Tree& t0, Tree& t1)		{ return isTree(t, SIGATTACH, t0, t1); 	}
 
-	 
+
 
 Tree addNums(Tree a, Tree b)
 {
@@ -262,13 +262,13 @@ Tree divExtendedNums(Tree a, Tree b)
 }
 
 Tree minusNum(Tree a)
-{ 
-	return tree(minusNode(a->node())); 
+{
+	return tree(minusNode(a->node()));
 }
 
 Tree inverseNum(Tree a)
-{ 
-	return tree(inverseNode(a->node())); 
+{
+	return tree(inverseNode(a->node()));
 }
 
 bool isSigAdd(Tree a, Tree& x, Tree& y)
@@ -309,7 +309,7 @@ Sym SIGTUPLEACCESS 	= symbol ("SigTupleAccess");
 Tree sigTuple (int mode, Tree ls)					{ return tree(SIGTUPLE, tree(mode), ls); }
 bool isSigTuple (Tree s, int* mode, Tree& ls)		{ Tree m; return isTree(s, SIGTUPLE, m, ls) && isInt(m->node(), mode); }
 
-// Access the components of a tuple. 
+// Access the components of a tuple.
 // ts is tuple of signals, idx is a scalar signal between 0..n
 Tree sigTupleAccess(Tree ts, Tree idx)				{ return tree(SIGTUPLEACCESS, ts, idx); }
 bool isSigTupleAccess(Tree s, Tree& ts, Tree& idx)	{ return isTree(s, SIGTUPLEACCESS, ts, idx); }
@@ -319,20 +319,20 @@ Tree sigCartesianProd (Tree s1, Tree s2)
 {
 	Tree 	l1, l2;
 	int		m1, m2;
-	
+
 	if (isSigTuple(s1, &m1, l1) && (m1 == 0)) {
 		// nothing to do
 	} else {
 		l1 = list1(s1);
 	}
-	
+
 	if (isSigTuple(s2, &m2, l2) && (m2 == 0)) {
 		// nothing to do
 	} else {
 		l2 = list1(s2);
 	}
-	
-	
+
+
 	return sigTuple(0, concat(l1,l2));
 }
